Migrate to same spec file
