+++ src/machines/modelingMachine.spec.ts
@@ -1,6 +1,11 @@
-/** Engine-using integration tests of modelingMachine.
- * For engineless unit tests, see modelingMachine.test.ts */
-import { assertParse, recast, type CallExpressionKw } from '@src/lang/wasm'
+/** Engine-using integration tests of modelingMachine. */
+import {
+  assertParse,
+  recast,
+  type Artifact,
+  type ArtifactGraph,
+  type CallExpressionKw,
+} from '@src/lang/wasm'
 import type { SceneGraphDelta } from '@rust/kcl-lib/bindings/FrontendApi'
 import { err } from '@src/lib/trap'
 import toast from 'react-hot-toast'
@@ -18,8 +23,11 @@ import { vi } from 'vitest'
 import { getConstraintInfoKw } from '@src/lang/std/sketch'
 import { ARG_END_ABSOLUTE, ARG_INTERIOR_ABSOLUTE } from '@src/lang/constants'
 import { removeSingleConstraintInfo } from '@src/lang/modifyAst'
-import { dummyInitSketchGraphDelta } from '@src/machines/modelingSharedContext'
-import { generateModelingMachineDefaultContext } from '@src/machines/modelingSharedContext'
+import {
+  dummyInitSketchGraphDelta,
+  generateModelingMachineDefaultContext,
+  modelingMachineInitialInternalContext,
+} from '@src/machines/modelingSharedContext'
 import {
   removeSingleConstraint,
   transformAstSketchLines,
@@ -43,6 +51,10 @@ let rustContextInThisFile: RustContext = null!
 let commandBarActorInThisFile: CommandBarActorType = null!
 let machineManagerInThisFile: MachineManager = null!
 
+const TESTS_WITHOUT_ENGINE_WORLD = [
+  'routes a cursor inside a sketch block segment to sketch solve edit',
+]
+
 /**
  * Every it test could build the world and connect to the engine but this is too resource intensive and will
  * spam engine connections.
@@ -50,6 +62,10 @@ let machineManagerInThisFile: MachineManager = null!
  * Reuse the world for this file. This is not the same as global singleton imports!
  */
 beforeEach(async () => {
+  const currentTestName = expect.getState().currentTestName ?? ''
+  if (TESTS_WITHOUT_ENGINE_WORLD.some((name) => currentTestName.includes(name)))
+    return
+
   if (instanceInThisFile) {
     return
   }
@@ -71,7 +87,7 @@ beforeEach(async () => {
 })
 
 afterAll(() => {
-  engineCommandManagerInThisFile.tearDown()
+  engineCommandManagerInThisFile?.tearDown()
 })
 
 describe('modelingMachine.test.ts', () => {
@@ -1509,6 +1525,115 @@ sketch001 = sketch(on = YZ) {
         expect(toastErrorSpy).toHaveBeenCalledWith(
           'Unable to enter sketch while KCL has parse errors.'
         )
+      })
+
+      it('routes a cursor inside a sketch block segment to sketch solve edit', async () => {
+        const pathToNode = [['body', '']] as any
+        const sketchBlockCodeRef = {
+          range: [0, 100, 0] as [number, number, number],
+          pathToNode,
+          nodePath: { steps: [] },
+        } as any
+        const segmentCodeRef = {
+          range: [35, 85, 0] as [number, number, number],
+          pathToNode,
+          nodePath: { steps: [] },
+        } as any
+        const sketchBlock: Extract<Artifact, { type: 'sketchBlock' }> = {
+          type: 'sketchBlock',
+          id: 'sketch-block-1',
+          codeRef: sketchBlockCodeRef,
+          planeId: 'plane-1',
+          sketchId: 7,
+        }
+        const path: Artifact = {
+          type: 'path',
+          subType: 'sketch',
+          id: 'path-1',
+          codeRef: sketchBlockCodeRef,
+          planeId: 'plane-1',
+          segIds: ['segment-1'],
+          trajectorySweepId: null,
+          consumed: false,
+          sketchBlockId: sketchBlock.id,
+        }
+        const segment: Artifact = {
+          type: 'segment',
+          id: 'segment-1',
+          pathId: path.id,
+          edgeIds: [],
+          commonSurfaceIds: [],
+          codeRef: segmentCodeRef,
+        }
+        const artifactGraph: ArtifactGraph = new Map()
+        artifactGraph.set(sketchBlock.id, sketchBlock)
+        artifactGraph.set(path.id, path)
+        artifactGraph.set(segment.id, segment)
+
+        const context = {
+          ...modelingMachineInitialInternalContext,
+          selectionRanges: {
+            graphSelections: [
+              {
+                codeRef: {
+                  range: [45, 45, 0],
+                  pathToNode,
+                },
+              },
+            ],
+            otherSelections: [],
+          },
+          kclManager: {
+            artifactGraph,
+            hidePlanes: vi.fn(),
+            showPlanes: vi.fn(),
+            sceneInfra: {
+              animate: vi.fn(),
+              resetMouseListeners: vi.fn(),
+              camControls: {
+                enablePan: true,
+                enableRotate: true,
+                syncDirection: 'engineToClient',
+              },
+            },
+          },
+          rustContext: {},
+          engineCommandManager: {},
+          wasmInstance: {},
+          commandBarActor: {},
+          machineManager: {},
+        } as any
+
+        let receivedInput: any
+        let actorStarted = false
+        const machine = modelingMachine.provide({
+          actors: {
+            'animate-to-existing-sketch-solve': fromPromise(
+              async ({ input }) => {
+                receivedInput = input
+                actorStarted = true
+                return await new Promise<any>(() => {})
+              }
+            ),
+          },
+        })
+
+        const actor = createActor(machine, { input: context }).start()
+
+        actor.send({ type: 'Enter sketch' })
+
+        await waitForCondition(() => actorStarted)
+
+        expect(actor.getSnapshot().value).toBe(
+          'animating to existing sketch solve'
+        )
+        expect(receivedInput).toEqual(
+          expect.objectContaining({
+            artifactId: sketchBlock.id,
+          })
+        )
+
+        actor.stop()
       })
     })
